zeroconf - a bit of renaming
